+++ toolkit/scaffold/project/cli.py
@@ -4,6 +4,7 @@ import sys
 
 import click
 
+from toolkit.scaffold.project.ansible.cli import create_ansible_project
 from toolkit.scaffold.project.cpp.cli import create_cpp_project
 from toolkit.scaffold.project.docker.cli import create_docker_project
 from toolkit.scaffold.project.golang.cli import create_golang_project
@@ -15,6 +16,8 @@ from toolkit.scaffold.project.python.cli import create_python_project
 def main():
     pass
 
+
+main.add_command(create_ansible_project, "ansible")
 
 main.add_command(create_cpp_project, "cpp")