+++ toolkit/template/scaffold/project/python/package/{{project_slug.snake_case}}/cli.py
@@ -1,5 +1,4 @@
 """Console script for {{project_slug.snake_case}}."""
-import os
 import sys
 
 import click