+++ HISTORY.md
@@ -4,6 +4,8 @@ v1.4.5 (2022-08-09)
 
 - Added
     - use multiple templates at the same time
+    - ansible project scaffold
+    - golang example and cli project scaffold
 
 v1.4.4 (2022-08-09)